backend/requirements.txt
# FastAPI 및 기본 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스 (MySQL → PostgreSQL)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# 인증 및 보안 (JWT 토큰 기반)
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# 이메일 인증
fastapi-mail==1.4.1

# 구글 OAuth 및 API
authlib==1.2.1
httpx==0.25.2
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# 유틸리티
python-dotenv==1.0.0
openpyxl==3.1.2
cryptography==41.0.7

# CORS 설정 (React 연동용)
fastapi-cors==0.0.6
